fix: redux thread bad set state error
