+++ src/components/generation/threads-chat-initializer.tsx
@@ -1,0 +1,18 @@
+"use client";
+import { useAppStore } from "@/hooks/redux";
+import { setNewThreadId } from "@/lib/redux/features/chat";
+import { useRef } from "react";
+
+export default function ThreadsChatInitializer({
+  threadId,
+}: {
+  threadId: string;
+}) {
+  const store = useAppStore();
+  const initialized = useRef(false);
+  if (!initialized.current) {
+    store.dispatch(setNewThreadId(threadId));
+    initialized.current = true;
+  }
+  return null;
+}
